sourcerer cleanups
