+++ x-pack/plugins/security_solution/public/common/components/inspect/modal.tsx
@@ -27,7 +27,8 @@ import { useLocation } from 'react-router-dom';
 import type { InputsModelId } from '../../store/inputs/constants';
 import { NO_ALERT_INDEX } from '../../../../common/constants';
 import * as i18n from './translations';
-import { getScopeFromPath, useSourcererDataView } from '../../../sourcerer/containers';
+import { getScopeFromPath } from '../../../sourcerer/containers/sourcerer_paths';
+import { useSourcererDataView } from '../../../sourcerer/containers';
 import { SourcererScopeName } from '../../../sourcerer/store/model';
 
 export interface ModalInspectProps {
